pyproject.toml
[project]
name = "mcp-maze-runner"
version = "0.1.0"
description = "An interactive maze game controlled through natural language chat, built to demonstrate the Model Context Protocol (MCP) integration with AI agents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
    "mcp>=1.21.0",
    "streamlit>=1.40.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
